Added files via upload
These python scripts were used for cleaning Illumina universal primers from short sequences of DNA for 16S rRNA sequencing for micobiome work. The original version reads a fastq file, the FNA version a fasta file.
Trim_to_amplicon_primers is designed to remove sequencing primers and other extraneous sequence that is outside the amplicon primers. The code has the DNA sequences for the forward and reverse primers hard-coded, so needs to be edited to change them.